flask_server/requirements.txt
Flask==3.0.0
Flask-SocketIO>=5.3.6
gevent-websocket==0.10.1
greenlet==3.0.0
redis